Update element_extension queries to use include rather than join
In the previous commits, to fix the issue on search for
searchable_contents, joins was removed. But because of this,
the associated models cannot be retrieved, since similar issue
is faced. So, for eg, it is not possible to use
element.searchable_essence_texts since, it would raise error.
To fix this, `element` has been included. This makes sure that
the table gets loaded and the fields are available.
The other option was to use references(:alchemy_elements) but it
raised similar error since the table wasn't pre-loaded.
Also, the current search works as expected, so this change
doesn't impact the current working