Remove nearly all migrations from dummy app
Migrate the dummy fresh with Alchemy 7.0 and remove nearly all migration. Ignore the migrate folder in the gitignore to store only the generated schema.rb. The only exception is the create_user migration that is a custom migration of the dummy app to have the possibility to run user specific feature specs.