Use an inner select for CanCanCan abilities
The previous solution only worked if the pg_search_document had a page_id. If another search document without a page relation was added and an ability was present, the document couldn't be found, because the generate SQL had an inner join that was using the page_id. The new solution is using a sub select and only filters pg_search_documents if the page_id is present on that document entries.